rename classes to models
