+++ app/src/main/java/com/example/habittracker/AddNewHabitActivity.java
@@ -12,7 +12,7 @@ import android.widget.ToggleButton;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.SwitchCompat;
 
-import com.example.habittracker.classes.Habit;
+import com.example.habittracker.models.Habit;
 import com.example.habittracker.controllers.HabitListController;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
