+++ app/src/main/java/com/example/habittracker/HabitsTodayFragment.java
@@ -1,6 +1,5 @@
 package com.example.habittracker;
 
-import android.os.Build;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -9,24 +8,16 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
-import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
 import androidx.fragment.app.Fragment;
 
 import com.example.habittracker.adapters.HabitListAdapter;
-import com.example.habittracker.classes.Habit;
-import com.example.habittracker.classes.HabitList;
+import com.example.habittracker.models.Habit;
+import com.example.habittracker.models.HabitList;
 import com.example.habittracker.controllers.HabitListController;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.FirebaseFirestoreException;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.Calendar;
-import java.util.Map;
 
 public class HabitsTodayFragment extends Fragment {
