+++ app/src/main/java/com/example/habittracker/HabitsTotalFragment.java
@@ -11,8 +11,8 @@ import android.widget.ListView;
 import androidx.fragment.app.Fragment;
 
 import com.example.habittracker.adapters.HabitListAdapter;
-import com.example.habittracker.classes.Habit;
-import com.example.habittracker.classes.HabitList;
+import com.example.habittracker.models.Habit;
+import com.example.habittracker.models.HabitList;
 import com.example.habittracker.controllers.HabitListController;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.firebase.auth.FirebaseAuth;
