+++ app/src/main/java/com/example/habittracker/controllers/HabitEventsController.java
@@ -2,18 +2,13 @@ package com.example.habittracker.controllers;
 
 import android.util.Log;
 
-import androidx.annotation.NonNull;
-
-import com.example.habittracker.classes.Habit;
-import com.example.habittracker.classes.HabitEvent;
-import com.example.habittracker.classes.HabitEventList;
-import com.example.habittracker.classes.HabitList;
+import com.example.habittracker.models.Habit;
+import com.example.habittracker.models.HabitEvent;
+import com.example.habittracker.models.HabitEventList;
+import com.example.habittracker.models.HabitList;
 import com.example.habittracker.interfaces.OnHabitListRetrieved;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
