+++ app/src/main/java/com/example/habittracker/models/HabitList.java
@@ -1,4 +1,4 @@
-package com.example.habittracker.classes;
+package com.example.habittracker.models;
 
 import android.os.Build;
