+++ app/src/main/java/com/example/habittracker/ui/following/FollowingViewModel.java
@@ -4,8 +4,7 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-import com.example.habittracker.classes.Habit;
-import com.example.habittracker.classes.User;
+import com.example.habittracker.models.User;
 
 import java.util.List;
